+++ R/discretise.R
@@ -27,20 +27,17 @@
 discretise <- function(m, by = 1, debug = 0) {
     mooringDebug(debug, "discretise() {\n")
     if (!is.mooring(m)) {
-        print(class(m))
         stop("only works for objects created by mooring()")
     }
     if (by <= 0) {
         stop("by must be a positive number")
     }
-    n <- length(m)
-    rval <- list()
-    class(rval) <- "mooring" # FIXME: this is tedious; perhaps switch to S4 system
+    elementsNew <- list() # copy elements into this, then make mooring from it
     group <- 1L
-    for (item in m) {
-        mooringDebug(debug, "  handling item of class c(\"", paste(class(item), collapse = "\", \""), "\")\n", sep = "")
-        if (inherits(item, "wire") || inherits(item, "chain")) {
-            height <- item$height
+    for (element in m@elements) {
+        mooringDebug(debug, "  handling item of class c(\"", paste(class(element), collapse = "\", \""), "\")\n", sep = "")
+        if (is.wire(element) || is.chain(element)) {
+            height <- element@height
             n <- as.integer(1 + floor(height / by))
             # message(    "INITIAL: height=", height, ", by=", by, ", n=", n)
             # Ensure at least 20 chunks
@@ -49,33 +46,31 @@ discretise <- function(m, by = 1, debug = 0) {
                 by <- height / n
                 # message("  LATER:", height, ", by=", by, ", n=", n)
             }
-            portion <- item
-            portion$height <- height / n
-            portion$area <- portion$area / n
-            portion$buoyancy <- portion$buoyancy / n
-            portion$group <- group # so we can undo this later
+            portion <- element
+            portion@height <- height / n
+            portion@area <- portion@area / n
+            portion@buoyancy <- portion@buoyancy / n
+            portion@group <- group # so we can undo this later
             for (i in seq_len(n)) {
-                rval[[1L + length(rval)]] <- portion
+                elementsNew[[1L + length(elementsNew)]] <- portion
             }
             group <- group + 1L
         } else {
-            rval[[1L + length(rval)]] <- item
+            elementsNew[[1L + length(elementsNew)]] <- element
         }
     }
-    nrval <- length(rval)
-    waterDepth <- rval[[nrval]]$depth
+    # FIXME: why is rev() required here? Is it an S7-change issue?
+    rval <- mooringS7(rev(elementsNew), waterDepth = m@waterDepth)
+    rval@elements <- rev(rval@elements)
     # Compute z and tau values. (Leave x values alone.)
-    # OLD z <- rev(-rval[[length(rval)]]$depth + cumsum(sapply(rval, function(x) x$height)))
+    # OLD z <- rev(-rval[[length(rval)]]$depth + cumsum(sapply(rval, \(x) x$height)))
     tau <- tension(rval, stagnant = TRUE) # FIXME: ok?
-    zz <- -waterDepth
-    for (i in rev(seq_along(rval))) {
-        zz <- zz + rval[[i]]$height
-        rval[[i]]$z <- zz # z is defined at TOP of item
-        rval[[i]]$tau <- tau[i]
+    zz <- -m@waterDepth
+    for (i in rev(seq_along(rval@elements))) {
+        zz <- zz + rval@elements[[i]]@height
+        rval@elements[[i]]@z <- zz # z is defined at TOP of item
+        rval@elements[[i]]@tau <- tau[i]
     }
-    class(rval) <- "mooring"
-    attr(rval, "discretised") <- TRUE
-    attr(rval, "waterDepth") <- attr(m, "waterDepth")
     mooringDebug(debug, "} # discretise()\n")
     rval
 } # discretise