+++ R/z.R
@@ -15,7 +15,7 @@
 #'
 #' @examples
 #' library(mooring)
-#' m <- mooring(anchor(depth = 120), wire(length = 100), float("HMB 20"))
+#' m <- mooring(anchor(), wire(length = 100), float("HMB 20"), waterDepth = 120)
 #' z(m)
 #'
 #' @export
@@ -23,18 +23,18 @@
 #' @author Dan Kelley
 z <- function(m, stagnant = FALSE, skipWire = FALSE) {
     e <- m@elements
-    message("in z 1")
+    # message("in z 1")
     rval <- if (stagnant) {
-        message("in z 2 (stagnant)")
+        # message("in z 2 (stagnant)")
         if (length(e[[1]]@z0)) {
-            message("in z 2 (stagnant) z0 is not empty")
+            # message("in z 2 (stagnant) z0 is not empty")
             sapply(e, \(ee) ee@z0)
         } else {
-            message("in z 2 (stagnant) z0 is empty, so using z")
+            # message("in z 2 (stagnant) z0 is empty, so using z")
             sapply(e, \(ee) ee@z)
         }
     } else {
-        message("in z 3 (not stagnant)")
+        # message("in z 3 (not stagnant)")
         sapply(e, \(ee) ee@z)
     }
     if (skipWire) rval[!is.wire(m)] else rval # FIXME: this is likely wrong